finance_app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl_cffi==0.11.4
dash==3.0.4
dash-bootstrap-components==2.0.3
Flask==3.0.3
frozendict==2.4.6
google-auth==2.40.3
google-genai==1.22.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multitasking==0.0.11
narwhals==1.44.0
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
peewee==3.18.1
platformdirs==4.3.8
plotly==6.1.2
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
retrying==1.4.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.0.6
yfinance==0.2.63
zipp==3.23.0
